% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_check_col.R
\name{coding_check_col}
\alias{coding_check_col}
\title{Check if words or regex appears in column}
\usage{
coding_check_col(list_words, column)
}
\arguments{
\item{list_words}{vector of words or regular expressions}

\item{column}{dataframe column (as vector)}
}
\value{
boolean vector (true / false)
}
\description{
This helper function lets you efficiently search for a list of words within a question, and re-code those responses.
For example, you could use this inside a dplyr::case_when() call to re-code any responses that mention the word 'test' into a 'Test' response group.
}
\examples{
coding_check_col(c('test', 'words'), c('this is test response 1', 'this is response 2', 'these are some words'))
}
