% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_create_orig.R
\name{coding_create_orig}
\alias{coding_create_orig}
\title{Extract only pre-set options from a multi-code question}
\usage{
coding_create_orig(column, manifest_pattern)
}
\arguments{
\item{column}{dataframe column (as vector)}

\item{manifest_pattern}{string with regular expression pattern}
}
\value{
string vector with comma-separated pre-set options
}
\description{
Where a column contains a "tick all that apply" type multi-code question, you want to treat the pre-set options just as they are.
This function allows you to extract just those responses, if you have a manifest pattern.
A manifest pattern is a vector of the pre-set options, prepared to have a regular expression shape.
You can do this using the coding_create_pattern function from this package.
}
\examples{
coding_create_orig(c('pre-set,manual', 'manual'), '(pre-set)')
}
