% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_create_other.R
\name{coding_create_other}
\alias{coding_create_other}
\title{Extract only manually entered text from a multi-code question}
\usage{
coding_create_other(column, manifest_pattern)
}
\arguments{
\item{column}{dataframe column (as vector)}

\item{manifest_pattern}{string with regular expression pattern}
}
\value{
string vector with comma-separated pre-set options
}
\description{
For multi-code "tick all that apply" type questions, there is sometimes an "Other, please specify" option.
This function extracts only this section, for grouping into categories.
}
\examples{
coding_create_other(c('pre-set,manual', 'manual'), '(pre-set)')
}
