% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_create_pattern.R
\name{coding_create_pattern}
\alias{coding_create_pattern}
\title{Create regex pattern from list of pre-set answer options}
\usage{
coding_create_pattern(manifest)
}
\arguments{
\item{manifest}{vector of pre-set answer options}
}
\value{
regular expression string
}
\description{
Create regex pattern from list of pre-set answer options
}
\examples{
coding_create_pattern(c('answer option 1', 'answer option 2'))
}
