% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_create_total.R
\name{coding_create_total}
\alias{coding_create_total}
\title{Combine pre-set and coded multi-code responses}
\usage{
coding_create_total(orig, other)
}
\arguments{
\item{orig}{string vector as produced by coding_create_orig()}

\item{other}{string vector as produced by coding_create_other()}
}
\value{
string vector
}
\description{
After extracting the original options (coding_create_orig) and grouping (coding_check_col) the other (coding_create_other) responses,
this function combines them together.
}
\examples{
coding_create_total(c('option 1,option 2'), c('typed option'))
}
