% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_qtypes.R
\name{detect_qtypes}
\alias{detect_qtypes}
\title{Detect question types}
\usage{
detect_qtypes(
  response_col,
  unique_vals = 12,
  split_perc = 0.5,
  prop_total = 0.85,
  prop_common = 0.05
)
}
\arguments{
\item{response_col}{column from response dataframe}
}
\value{

}
\description{
Looks at the possible answers in a column, and determines which type of question
is most likely. This is based on rules of thumb and can be adapted for different
uses.
}
