% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_grouping_var.R
\name{prep_grouping_var}
\alias{prep_grouping_var}
\title{Set grouping variable to tidyeval}
\usage{
prep_grouping_var(grouping_var)
}
\arguments{
\item{grouping_var}{with quotation marks (string) or without (tidyeval)}
}
\value{
value prepared for tidy evaluation
}
\description{
Set grouping variable to tidyeval
}
\examples{
prep_grouping_var('column_1')
}
