% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types.R
\name{question_types}
\alias{question_types}
\title{Group columns into question types}
\usage{
question_types(dummy_response)
}
\arguments{
\item{dummy_response}{}
}
\value{
three-element list
}
\description{
Returns a list containing vectors with column names for categorical, multi-choice, and free text questions.
}
\examples{
question_types(dummy_response)
}
