% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_question_types.R
\name{survey_question_types}
\alias{survey_question_types}
\title{Group columns into question types}
\usage{
survey_question_types(dummy_response)
}
\arguments{
\item{dummy_response}{dataframe}
}
\value{
three-element list
}
\description{
Returns a list containing vectors with column names for categorical, multi-choice, and free text questions.
}
\examples{
survey_question_types(dummy_response)
}
