% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_response_barplot.R
\name{survey_response_barplot}
\alias{survey_response_barplot}
\title{Visualise responses}
\usage{
survey_response_barplot(...)
}
\arguments{
\item{survey_response}{list}
}
\value{
A ggplot2 object
}
\description{
Take a consultation response spreadsheet (as transformed by Hmisc::describe), and turn it into a barplot.
}
\examples{
survey_response_barplot(Hmisc::describe(dummy_response))
}
