% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_response_tables.R
\name{survey_response_tables}
\alias{survey_response_tables}
\title{Extract and clean responses}
\usage{
survey_response_tables(dummy_response, qtypes, min_n = 5)
}
\arguments{
\item{dummy_response}{dataframe}

\item{qtypes}{list with elements categorical, multichoice, and freetext containing vectors of column names}

\item{min_n}{smallest group of answers allowed (statistical disclosure limit)}
}
\value{
list of tibbles
}
\description{
Take a consultation response spreadsheet and turn it into a list of neat tibbles of values and frequencies.
}
\examples{
survey_response_tables(dummy_response, survey_question_types(dummy_response))
}
