% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_text.R
\name{survey_summary_text}
\alias{survey_summary_text}
\title{Summarise categorical responses as text}
\usage{
survey_summary_text(survey_response)
}
\arguments{
\item{survey_response}{list}
}
\value{
glue string
}
\description{
Take a question frequency table (as produced by response_tables), and summarise the responses as three-bullet text.
Requires the frequency table to have 'Response', 'Frequency', and 'Percentage' columns.
To apply to the full output of response_tables, use lapply.
}
\examples{
lapply(survey_response_tables(dummy_response, qtypes = survey_question_types(dummy_response)), survey_summary_text)
}
