% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_drop_keyword_responses.R
\name{text_drop_keyword_responses}
\alias{text_drop_keyword_responses}
\title{Remove responses with keywords}
\usage{
text_drop_keyword_responses(data, text_col, keywords = c("insert_keywords"))
}
\arguments{
\item{data}{data.frame containing all responses}

\item{text_col}{column containing free text}

\item{keywords}{responses containing any of these words will be removed. Not case-sensitive.}
}
\value{
data.frame with responses that match keywords removed
}
\description{
This function may be useful if you want to remove responses which mention a particular word or phrase.
Note: this removes the whole response, not only the word itself - if you want to
remove one word only, use the text_remove_words() function.
}
\examples{
text_drop_keyword_responses(dummy_response, colnames(dummy_response)[7], 'malesuada')
}
