% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_drop_multiple_allcaps_responses.R
\name{text_drop_multiple_allcaps_responses}
\alias{text_drop_multiple_allcaps_responses}
\title{Remove responses with multiple all-caps words}
\usage{
text_drop_multiple_allcaps_responses(
  data,
  text_col,
  max_cap = 3,
  ignore_caps = c("NHS")
)
}
\arguments{
\item{data}{data.frame containing all responses}

\item{text_col}{column containing free text}

\item{max_cap}{maximum number of all-caps words to allow per response}

\item{ignore_caps}{string vector of all-caps words to ignore in this count (eg. legitimate acronyms)}
}
\value{
data.frame of responses with all-caps responses filtered out
}
\description{
In some cases, you don't want to include responses that are written fully capitalised.
However, you must be careful to ignore legitimate all-caps words, such as acronyms in the domain of the consultation (eg. 'NHS').
}
\examples{
text_drop_multiple_allcaps_responses(data.frame(doc_id = c(1, 2), text = c('TEST THE FUNCTION', 'TEST NHS FUNCTION')), text_col = 'text', max_cap = 2)
}
