% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_dtm_prep.R
\name{text_dtm_prep}
\alias{text_dtm_prep}
\title{Prepare data for topic modelling}
\usage{
text_dtm_prep(unnest_data, grouping_var, word_col = "word")
}
\arguments{
\item{unnest_data}{dataframe with unnested tokens}

\item{grouping_var}{column used for determining what consistutes a document; with quotation marks (string)
If each response is a row, add a column with row identifiers to be the grouping_var.}

\item{word_col}{column name containing the unnested tokens}
}
\value{
List containing the prepared data and a Document-Term Matrix
}
\description{
Calculate TF-IDF scores and create a Document-Term Matrix from your dataset.
}
\examples{
tidytext::unnest_tokens(dummy_response, 'word', colnames(dummy_response)[7]) \%>\%
text_dtm_prep(., 'response_id')


}
