% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stem_words.R
\name{text_stem_words}
\alias{text_stem_words}
\title{Stem words before analysis}
\usage{
text_stem_words(unnest_data, word_col = "word")
}
\arguments{
\item{unnest_data}{dataframe with unnested free text data (a row per word, eg. as prepared by tidytext::unnest_tokens)}

\item{word_col}{column name containing word tokens}
}
\value{
dataframe with words replaced with their stemmed version
}
\description{
Stem words before analysis
}
\examples{
text_stem_words(data.frame(doc_id = c(1, 2, 3, 4), word = c('test', 'testing', 'tester', 'word')), word_col = 'word')
}
