% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tf_idf_out.R
\name{text_tf_idf_out}
\alias{text_tf_idf_out}
\title{Calculate TF-IDF scores}
\usage{
text_tf_idf_out(unnest_data, grouping_var)
}
\arguments{
\item{grouping_var}{column that groups each row. May be with (string) or without (tidy evaluation) quotation marks}

\item{data}{pre-cleaned response data}
}
\value{
dataframe with a row for each group-word combination, and the tf-idf score
}
\description{
Calculate TF-IDF scores
}
\examples{
tidytext::unnest_tokens(dummy_response, 'word', colnames(dummy_response)[7]) \%>\%
text_tf_idf_out(., colnames(dummy_response)[1])


}
