% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_unnest_remove_stem_words.R
\name{text_unnest_remove_stem_words}
\alias{text_unnest_remove_stem_words}
\title{Pre-clean consultation responses}
\usage{
text_unnest_remove_stem_words(
  data,
  text_col,
  token = "words",
  custom_words = c(""),
  ...
)
}
\arguments{
\item{data}{dataframe of responses}

\item{text_col}{name of column containing free text to be analysed}

\item{token}{what text entity to be analysed? by default word-by-word, but other options are defined in tidytext::unnest_tokens}
}
\value{
dataframe with stemmed, unnested words with removed words
}
\description{
Takes a dataframe with a free-text column as input. Unnests words from the free text, removes stopwords, profanity, and number-only words,
removes an optional custom list of words, and stems words (using hunspell) so they can be grouped better.
}
\details{
Note that, although you may choose a token other than "words" (like "sentences"), the word removal and stemming expects word tokens only.
}
\examples{
text_unnest_remove_stem_words(dummy_response, colnames(dummy_response)[7], "words")
}
