% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_vis_tf_idf.R
\name{text_vis_tf_idf}
\alias{text_vis_tf_idf}
\title{Visualise tf-idf output}
\usage{
text_vis_tf_idf(tfidf_data, grouping_var, top_n = 10)
}
\arguments{
\item{tfidf_data}{dataframe as produced by text_tf_idf_out function}

\item{grouping_var}{a column name, with quotes (string) or without (tidyeval)}

\item{top_n}{Top x number of words by tf-idf score}
}
\value{
ggplot2 plot
}
\description{
Visualise tf-idf output
}
\examples{
text_unnest_remove_stem_words(dummy_response, colnames(dummy_response)[7], "words") \%>\%
text_tf_idf_out(., colnames(dummy_response)[1]) \%>\%
text_vis_tf_idf(., colnames(dummy_response)[1])

}
