% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_stem_and_remove_stopwords.R
\name{unnest_stem_and_remove_stopwords}
\alias{unnest_stem_and_remove_stopwords}
\title{Pre-clean consultation responses}
\usage{
unnest_stem_and_remove_stopwords(data, text_col, token = "words")
}
\arguments{
\item{data}{dataframe of responses}

\item{text_col}{column name containing free text to be analysed}

\item{token}{what text entity to be analysed? by default word-by-word, but other options are defined in tidytext::unnest_tokens}
}
\value{
dataframe with stemmed, unnested words with removed stopwords
}
\description{
Pre-clean consultation responses
}
\examples{
unnest_stem_and_remove_stopwords(dummy_response, colnames(dummy_response)[7], "words")
}
