% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_tables.R
\name{response_tables}
\alias{response_tables}
\title{Extract and clean responses}
\usage{
response_tables(dummy_response, qtypes, min_n = 5)
}
\arguments{
\item{dummy_response}{dataframe}
}
\value{
list of tibbles
}
\description{
Take a consultation response spreadsheet and turn it into a list of neat tibbles of values and frequencies.
}
\examples{
response_tables(dummy_response, question_types(dummy_response))
}
