% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_tables.R
\name{response_tables}
\alias{response_tables}
\title{Extract and clean responses}
\usage{
response_tables(...)
}
\arguments{
\item{survey_response}{list produced by Hmisc::describe}
}
\value{
tibble
}
\description{
Take a consultation response spreadsheet (as transformed by Hmisc::describe), and turn it into a neat tibble of values and frequencies.
}
\examples{
purrr::map(Hmisc::describe(dummy_response), response_tables)
}
