% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_text.R
\name{summary_text}
\alias{summary_text}
\title{Summarise categorical responses as text}
\usage{
summary_text(survey_response)
}
\arguments{
\item{survey_response}{list}
}
\value{
string
}
\description{
Take a consultation response spreadsheet (as transformed by Hmisc::describe), and summarise the responses as three-bullet text.
}
\examples{
summary_text(Hmisc::describe(dummy_response))
}
