% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_response_barplot.R
\name{survey_response_barplot}
\alias{survey_response_barplot}
\title{Visualise responses}
\usage{
survey_response_barplot(dummy_response)
}
\arguments{
\item{dummy_response}{dataframe}
}
\value{
list of ggplot2 objects
}
\description{
Take a consultation response spreadsheet, and turn it into a barplot for each column.
This uses the question_types() function to clean data first, including aggregating small categories and detecting free text columns.
It then makes barplots on the basis of the frequency tables produced by response_tables().
}
\examples{
survey_response_barplot(dummy_response)
}
