% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_best_n_topics.R
\name{text_best_n_topics}
\alias{text_best_n_topics}
\title{Select best number of topics for LDA}
\usage{
text_best_n_topics(data_dtm, k_opts)
}
\arguments{
\item{data_dtm}{Document-Term matrix as prepared by \code{\link{text_dtm_prep}}}

\item{k_opts}{integer vector of numbers of topics to be considered}
}
\value{
single integer: best number of topics
}
\description{
Select best number of topics for LDA
}
\examples{
text_dtm_prep(dummy_response, colnames(dummy_response)[7], 'response_id') \%>\%
text_best_n_topics(., seq(5, 10))
}
