% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_dtm_prep.R
\name{text_dtm_prep}
\alias{text_dtm_prep}
\title{Prepare data for topic modelling}
\usage{
text_dtm_prep(data, text_col, grouping_var)
}
\arguments{
\item{data}{dataframe containing responses}

\item{text_col}{column name containing free text}

\item{grouping_var}{column used for determining what consistutes a document; with quotation marks (string) or without (tidyeval)}
}
\value{
Document-Term Matrix from responses dataframe
}
\description{
Create a Document-Term Matrix from your dataset.
Includes text preparation (unnesting, stopword removal, and stemming).
Also includes creation of tf-idf transformation.
}
\examples{
text_dtm_prep(dummy_response, colnames(dummy_response)[7], colnames(dummy_response)[1])
}
