% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_dtm_prep.R
\name{text_dtm_prep}
\alias{text_dtm_prep}
\title{Prepare data for topic modelling}
\usage{
text_dtm_prep(data, text_col, grouping_var)
}
\arguments{
\item{data}{dataframe containing responses}

\item{text_col}{column name containing free text}

\item{grouping_var}{column used for determining what consistutes a document; with quotation marks (string).
If each response is a row, add a column with row identifiers to be the grouping_var.}
}
\value{
Document-Term Matrix from responses dataframe
}
\description{
Create a Document-Term Matrix from your dataset.
Includes text preparation (unnesting, stopword removal, and stemming).
Also includes creation of tf-idf transformation.
}
\examples{
dummy_response \%>\%
dplyr::mutate(response_id = row.names(.)) \%>\%
text_dtm_prep(., colnames(dummy_response)[7], 'response_id')
}
