% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_lda_dtm.R
\name{text_lda_dtm}
\alias{text_lda_dtm}
\title{Calculate LDA topic model}
\usage{
text_lda_dtm(data_dtm, k = 2, burnin = 1000, iter = 1000, keep = 50)
}
\arguments{
\item{data_dtm}{Document-Term matrix as prepared by text_dtm_prep}

\item{k}{Number of topics to detect}

\item{burnin}{parameter for Gibbs LDA}

\item{iter}{parameter for Gibbs LDA}

\item{keep}{parameter for Gibbs LDA}
}
\value{
list with elements: lda_out (full LDA topicmodel), beta (beta scores),
gamma (gamma scores), and logLik (Log likelihood of topicmodel object)
}
\description{
Calculate LDA topic model
}
\examples{
dummy_response \%>\%
dplyr::mutate(response_id = row.names(.)) \%>\%
text_dtm_prep(., colnames(dummy_response)[7], 'response_id') \%>\%
text_lda_dtm(., k = 5)
}
