% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tf_idf_out.R
\name{text_tf_idf_out}
\alias{text_tf_idf_out}
\title{Calculate TF-IDF scores}
\usage{
text_tf_idf_out(data, grouping_var)
}
\arguments{
\item{data}{pre-cleaned response data}

\item{grouping_var}{column that groups each row. May be with (string) or without (tidy evaluation) quotation marks}
}
\value{
dataframe with a row for each group-word combination, and the tf-idf score
}
\description{
Calculate TF-IDF scores
}
\examples{
text_tf_idf_out(text_unnest_stem_and_remove_stopwords(dummy_response, colnames(dummy_response)[7], "words"), colnames(dummy_response)[1])
}
